Consistently use u64 for addresses, even in tables

+++ storage_node/src/kv2/items/itemrecover_v.rs
@@ -29,7 +29,7 @@ pub(super) open spec fn item_recoverable<I>(
     where
         I: PmCopy,
 {
-    &&& sm.table.validate_row_addr(addr as int)
+    &&& sm.table.validate_row_addr(addr)
     &&& recover_object::<I>(s, addr + sm.row_item_start, addr + sm.row_item_crc_start) is Some
 }
